refactor(christmas): 방문 날짜 toInt() 적용

+++ src/main/kotlin/christmas/view/InputView.kt
@@ -9,7 +9,7 @@ class InputView {
     fun readDate(): VisitDate {
         println("12월 중 식당 예상 방문 날짜는 언제인가요? (숫자만 입력해 주세요!)")
         val input = Console.readLine()
-        return VisitDate(input)
+        return VisitDate(input.toInt())
     }
 
     fun readMenu(): Order {
